FEM: Add comperator and cout support for vectors
For future tests of FemConstraint a few comperators and std::cout
support was added to the FCTest library. The operators can compare
opencascade vectors with FreeCADs vectors directely with ==. This is
handy for google tests ASSERT_EQ and ASSERT_NE macros. The output
support enables google test to print better debug messeges when a test
fails (it displays the value of the two operands).