FEM: Fix build system so it works on ubuntu
Ubuntu doesn't have a package containing the shared libraries of
googletest. Instead of linking an external googletest it is now included
in the FreeCAD source as a git submodule (src/3rdParty/googletest) and
build together with the rest of FreeCAD. No additional dependencies are
nessesary to build FreeCAD with tests.

The newer find_package(FindPython) was used for the test code. FreeCAD
uses the older FindPythonLibs. Because FindPython doens't work on ubuntu
and to be consistent FindPythonLibs is now used instead of FindPython.