FEM: Add unittests testing simulation results
The added tests execute a predefined simulation and than check the
results against calculated values of a analytical approximation. Makes
heavy use of the test support library (support.py). Predifined cases are
stored in normal .FCStd files. Each unittests that adds a solver,
configures the settings and executes the simulation. Most of the work is
done by support.py though.